% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiBRM.R
\name{coef.semiBRM}
\alias{coef.semiBRM}
\title{Extract parameter estimates}
\usage{
\method{coef}{semiBRM}(object, ...)
}
\arguments{
\item{object}{a fitted 'semiBRM' object.}

\item{...}{further arguments (currently ignored).}
}
\value{
a numeric vector of parameter estimates.
}
\description{
This returns parameter estimates of the fitted semiparametric binary response model.
}
