% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GaussianKerNonpar}
\alias{GaussianKerNonpar}
\title{Gaussian Kernel Nonparametric Conditional Expectation}
\usage{
GaussianKerNonpar(Y, X, args, H, n_cores = 3L)
}
\arguments{
\item{Y}{A numeric vector of outcome variable}

\item{X}{A numeric vector or matrix of explanatory variable(s)}

\item{args}{A numeric vector or matrix of arguments at which nonparametric conditional
expectations are calculated}

\item{H}{A numeric vector of bandwidth(s)}

\item{n_cores}{the number of threads to parallize the computation of nonparametric
conditional expectation over \code{args}}
}
\value{
A numeric vector of nonparametric conditional estimates evaluated at \code{args}
}
\description{
This calculates the Gaussian kernel nonparametric conditional expectation
for given data points using \code{Rcpp} with \code{OpenMP}.
}
