% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GaussianKerNonparLeaveOneOut}
\alias{GaussianKerNonparLeaveOneOut}
\title{Leave-One-Out Gaussian Kernel Nonparametric Conditional Expectation}
\usage{
GaussianKerNonparLeaveOneOut(Y, X, H, n_cores = 3L)
}
\arguments{
\item{Y}{A numeric vector of outcome variable}

\item{X}{A numeric vector or matrix of explanatory variable(s)}

\item{H}{A numeric vector of bandwidth(s)}

\item{n_cores}{the number of threads to parallize the computation of nonparametric
conditional expectation over \code{X}}
}
\value{
A numeric vector of leave-one-out nonparametric conditional estimates evaluated at \code{X}
}
\description{
This calculates the leave-one-out Gaussian kernel nonparametric conditional
expectation over data points of explanatory variables using \code{Rcpp} with \code{OpenMP}.
}
