% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiBRM.R
\name{get_num_threads}
\alias{get_num_threads}
\title{Get the number of threads}
\usage{
get_num_threads()
}
\value{
an integer indicating the number of the threads set for parallelization.
}
\description{
This returns the number of threads set to parallelize computation of nonparametric conditional expectation.
}
