% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-func.R
\name{getInteriorPoints}
\alias{getInteriorPoints}
\title{Interior Points Indicator}
\usage{
getInteriorPoints(X, f)
}
\arguments{
\item{X}{a vector or matrix of numeric values}

\item{f}{a numeric number indicating cut-offs at \code{f} and 1-\code{f} percentiles}
}
\value{
An integer vector consisting only of ones and zeros.

If \code{X} is a matrix, interior points are searched on each column, and
the element-wise product of the interior point indicators among the columns are returned.
}
\description{
This returns an integer vector whose i-th element takes 1L if the i-th element
of \code{X} is an interior point lying between \code{f} and 1-\code{f} percentiles
and 0L otherwise.
}
\examples{
n <- 1000L
X <- rnorm(n)
f <- 0.025

# construct interior-point indicator from scratch
interior0 <- integer(n)
interior0[X > quantile(X, f) & X < quantile(X, 1-f)] <- 1L

# using the function
interior1 <- getInteriorPoints(X, f)

# compare outcome
all.equal(interior0, interior1) # TRUE
}
