% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiBRM.R
\name{predict.semiBRM}
\alias{predict.semiBRM}
\title{Predict semiparametric conditional probability}
\usage{
\method{predict}{semiBRM}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a fitted 'semiBRM' object.}

\item{newdata}{a data.frame or matrix at which conditional probabilities are computed.}

\item{...}{further arguments (currently ignored).}
}
\value{
a list with components:\tabular{ll}{
   prob \tab predictions. \cr
   non.endpoint \tab taking \code{TRUE} if estimated probabilities are evaluated at non-endpoints and \code{FALSE} otherwise. \cr
}
}
\description{
This calculates pointwise semiparametric conditional probability \code{Pr(y=1|x)} based
on the fitted semiparametric binary response model. If new data are not given, it returns
in-sample conditional probabilities using \code{\link{fitted.semiBRM}}.
}
