% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiBRM.R
\docType{package}
\name{semiBRM-package}
\alias{semiBRM-package}
\title{Semiparametric Binary Response Model}
\description{
Implementation of semiparametric binary response models theorized in
\insertCite{klein1993efficient;textual}{semiBRM}.
}
\details{
This package offers an implementation of semiparametric binary response models, theorized in
a seminal work in the semiparametric econometrics literature, \insertCite{klein1993efficient;textual}{semiBRM}.

Compared to other related packages that help run non-/semi-parametric analysis,
this relfects the econometrician's perspective to the best, taking conditions
for asymptotic properties seriously. For instance, the default choice of bandwidth size in
the Nadaraya-Watson estimator meets the conditions for square-root N consistency of the coefficient
estimator.

In turn, this package will be useful in conducing econometric analysis on binary response models.
For example, this package offers computation of marginal effects, which are often the most
important quantity of interest in binary response models in the economics context.

This is built upon \code{\link[Rcpp:Rcpp-package]{Rcpp}} \insertCite{eddelbuettel2011rcpp}{semiBRM} along with
\code{OpenMP}, speeding up computation of nonparametric conditional expectation over data points.
In author's opinion, the only significant disadvantage of this semiparametric approach to binary
choice models over 'standard' ones such as Probit and Logistic, particularly in the cross-sectional
setup, is high computation costs. In this package, this limitation is meaningfully overcome by
taking advantage of multithreading via \code{OpenMP} in \code{\link[Rcpp:Rcpp-package]{Rcpp}}.

The econometric theory underlying this package mostly comes from \insertCite{klein1993efficient;textual}{semiBRM}.
However, a few important parts are based on lectures of Prof. Klein at Rutgers University who
wrote that paper. Since \insertCite{klein1993efficient;textual}{semiBRM} was published,
he made several improvements in asymptotic theories. For example, the 'current version' of
asymptotics no longer requires the use of 'higher-order' kernel in the Nadaraya-Watson estimator,
which was asked in the original paper.
}
\references{
\insertAllCited{}
}
\author{
Hyungil Kye (Henry) \code{<}\email{hkye@economics.rutgers.edu}\code{>}
}
