% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiBRM.R
\name{set_num_threads}
\alias{set_num_threads}
\title{Set the number of threads}
\usage{
set_num_threads(x = parallel::detectCores() - 1L)
}
\arguments{
\item{x}{an integer indicating the number of the threads put in place for parallelization.}
}
\description{
This sets the number of threads to parallelize computation of nonparametric conditional expectation.
}
