% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiBRM.R
\name{vcov.semiBRM}
\alias{vcov.semiBRM}
\title{Calculate asymptotic variance-covariance matrix}
\usage{
\method{vcov}{semiBRM}(object, ...)
}
\arguments{
\item{object}{a fitted 'semiBRM' object.}

\item{...}{further arguments (currently ignored).}
}
\value{
a numeric matrix of the asymptotic variance-covariance.
}
\description{
This calculates the asymptotic variance-covariance matrix of the estimates in the
fitted semiparametric binary response model as the inverse of the negative hessian matrix.
}
