% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-func.R
\name{GaussianNadarayaWatsonEstimator}
\alias{GaussianNadarayaWatsonEstimator}
\title{Nadaraya-Watson estimator with Gaussian kernel}
\usage{
GaussianNadarayaWatsonEstimator(x, y, h, args = NULL)
}
\arguments{
\item{x}{a numeric vector or matrix of explanatory variable(s).}

\item{y}{a numeric vector of outcome variable.}

\item{h}{a numeric vector indicating bandwidth size(s) for each explanatory variable in \code{x}.}

\item{args}{a numeric vector or matrix of arguments at which nonparametric conditional
expectations are evaluated.}
}
\value{
a numeric vector of nonparametric estimates of conditional expectation \code{E[y|x=args]}.
}
\description{
This estimates nonparametrically conditional expectation \code{E[y|x=args]} using
the Nadaraya-Watson estimator. If target points \code{args} are not given, it will return
the leave-one-out version.
}
