% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiBRM.R
\name{predict.semiBRM}
\alias{predict.semiBRM}
\title{Predict semiparametric conditional probability}
\usage{
\method{predict}{semiBRM}(
  object,
  newdata = NULL,
  boot.se = FALSE,
  ci.level = 0.95,
  nboot = 300L,
  h = NULL,
  ...
)
}
\arguments{
\item{object}{a fitted 'semiBRM' object.}

\item{newdata}{a data.frame or matrix at which conditional probabilities are computed.}

\item{boot.se}{a logical indicating whether to report standard errors and confidence intervals.
If \code{boot.se = TRUE}, it calculates standard errors pointwise from semiparametric bootstrapping.}

\item{ci.level}{a numeric representing bootstrap confidence intervals. This is useful only when
\code{boot.se = TRUE}.}

\item{nboot}{an integer indicating the number of bootstrap replications. This is useful only when
\code{boot.se = TRUE}.}

\item{h}{a numeric of bandwidth size in the Nadaraya-Watson estimator. If not given, it will use
the Silverman's rule of thumb bandwidth, \code{h = sd(x)*1.06*N^(-1/5)}.}

\item{...}{further arguments (currently ignored).}
}
\value{
If \code{boot.se = FALSE}, then it will return a list with the following components:\tabular{ll}{
   prob \tab predictions. \cr
   non.endpoint \tab taking \code{TRUE} if estimated probabilities are evaluated at points away from boundaries and \code{FALSE} otherwise. \cr
}


If \code{boot.se = TRUE}, then it will return a list with the following components:\tabular{ll}{
   prob \tab predictions. \cr
   boot.se \tab semiparametric bootstrap standard errors. \cr
   boot.ci \tab semiparametric bootstrap confidence intervals. \cr
   non.endpoint \tab taking \code{TRUE} if estimated probabilities are evaluated at points away from boundaries and \code{FALSE} otherwise. \cr
}
}
\description{
This calculates pointwise semiparametric conditional probability \code{Pr(y=1|x)} based
on the fitted semiparametric binary response model. If new data are not given, it returns
in-sample conditional probabilities using \code{\link{fitted.semiBRM}}.
}
