% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiBRM.R
\name{semiBRM}
\alias{semiBRM}
\alias{semiBRM.default}
\alias{semiBRM.formula}
\title{Semiparametric binary response model: Parameter estimation}
\usage{
semiBRM(x, y, ...)

\method{semiBRM}{default}(x, y, r = 1/6.01, tau = 0.025, ...)

\method{semiBRM}{formula}(formula, data, r = 1/6.01, tau = 0.025, ...)
}
\arguments{
\item{x}{a numeric matrix of explanatory variables.}

\item{y}{a vector of integer, numeric, or factor of binary response outcomes, taking either 1 or 0 only.}

\item{...}{further arguments in "\code{\link[maxLik:maxLik]{maxLik::maxLik}}" such as \code{control}.}

\item{r}{a numeric number that controls the size of Silverman's rule-of-thumb bandwidth,
\code{h = sd(x)*N^(-r)}.}

\item{tau}{a numeric indicating cut-off levels for trimming in
\code{\link[=TrimmingIndicator]{TrimmingIndicator}(X,f)}, which assigns 1L to the
values in \code{X} lying between \code{tau}*100 and (1-\code{tau})*100 -th
percentiles and 0L to those outside this range.}

\item{formula}{a formula describing the model to be fitted..}

\item{data}{a data.frame containing variables in \code{formula}.}
}
\value{
object of class 'semiBRM' similar to that of 'maxLik' with elements:
\itemize{
\item \code{estimate}: estimated parameter values.
\item \code{log.likelihood}: log likelihood at the estimates.
\item \code{gradient}: a gradient vector at the estimates.
\item \code{hessian}: a hessian matrix at the estimates.
\item \code{code}: return code as detailed in \code{\link[maxLik:maxLik]{maxLik::maxLik}}.
\item \code{message}: a short message describing return code.
\item \code{iter}: the number of iterations performed for numerical optimization.
\item \code{control}: the optimization control parameters as detailed in
\code{\link[maxLik:maxLik]{maxLik::maxLik}}.
\item \code{model}: the model frame.
\item \code{r}: the bandwidth parameter for Silverman's rule-of-thumb bandwidth.
\item \code{trimming.level}: the trimming cutoff level, which is \code{tau} in function argument.
\item \code{call}: the matched call.
\item \code{formula}: the formula entered for estimation.
}
}
\description{
This implements quasi maximum likelihood estimation for parameters in semiparametric binary
response models.
}
\details{
This is the main function in the pacakge that performs parameter estimation of
semiparametric binary response models. It can take as arguments either matrix \code{x} and
vector \code{y} or \code{formula} and \code{data} to run estimation (see Examples below). The
default setup is reasonably good so that simply offering \code{x} and \code{y} or
\code{formula} and \code{data} would be enough in many cases.

Currently, only a single index model for binary outcome \code{y} is allowed. Importantly,
the first explanatory variable should be the one whose coefficient is strongly believed to be
different from zero, as the coefficients of other variables will be rescaled by that
of the first explanatory variable in estimation. This rescaling is unavoidable
in semiparametric approaches while it has "no" impact on estimation of conditional probability,
\code{Pr{y=1|x}}. The parameter estimates are found in quasi maximum likelihood estimation
using \code{\link[maxLik:maxLik]{maxLik::maxLik}} with BFGS method in place.

The theory in \insertCite{klein1993efficient;textual}{semiBRM} needs a well-defined trimming
indicator of 'index' that trims out boundary points to ensure a compact support. For it,
this runs estimation twice, as recommended in the paper, with the first estimation as 'pilot'
version and with the second one as the primary one. In the pilot version, the initial trimming
indicator is generated based on the original set of explanatory variables \code{x}, dropping out
observations near boundaries in any of the explanatory variables \code{x}, where observations
lying outside \code{[}\code{trimming.level*100}, \code{(1-trimming.level)*100}\verb{]} percentiles are
considered as being near boundaries, with the default value \code{trimming.level = 0.025}. Then,
the coefficient estimates of the pilot version are used to form the index, the linear combination
of explanatory variables with the estimated coefficients from the pilot version. Finally,
the trimming indicator for the primary version is generated from it at \code{trimming.level} and
taken to the log-likelihood function for parameter estimation.

The Silverman's rule of thumb bandwidth is put in place for the Nadaraya-Watson estimator, which
computes conditional probabilities. The bandwidth size is controlled by \code{r} with default value
\code{r = 6.01}, which satisfies conditions for consistency and asymptotic normality.

The package deploys OpenMP here, parallelizing computation of the Nadaraya-Watson estimator over
data points. The default value of the number of threads is \code{parallel::detectCores()-1L}.
To change it manually, please use \code{\link{set_num_threads}(x)}. Note that this affects all
functions that employ the Nadaraya-Watson estimator in the package. If set to be 1L,
multithreading will not be used.
}
\examples{
# data generating process
N <- 500L
X1 <- rnorm(N)
X2 <- (X1 + 2*rnorm(N))/sqrt(5) + 1
X3 <- rnorm(N)^2/sqrt(2)
X <- cbind(X1, X2, X3)
beta <- c(2, 2, -1, -1)
V <- as.vector(cbind(X, 1)\%*\%beta)
Y <- ifelse(V >= rnorm(N), 1L, 0L)

# identifiable set of parameters
ests_true <- c(1, -.5)

# using matrix/vector
qmle0 <- semiBRM(x = X, y = Y, control = list(iterlim=50))

# using formula and data
data <- data.frame(Y, X1, X2, X3)
qmle1 <- semiBRM(Y~X1+X2+X3, data = data, control = list(iterlim=50))

}
\references{
\insertAllCited{}
}
\seealso{
\link{GaussianNadarayaWatsonEstimator}, \link{TrimmingIndicator}
}
