% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiBRM.R
\name{semiBRM}
\alias{semiBRM}
\title{Quasi maximum likelihood estimation}
\usage{
semiBRM(
  X,
  Y,
  r = 1/7.1,
  cutoff.level = 0.025,
  num.threads = parallel::detectCores() - 1L,
  ...
)
}
\arguments{
\item{X}{a matrix of explanatory variables.}

\item{Y}{an integer vector of binary response outcome}

\item{r}{a numeric number that controls the size of Silverman's rule-of-thumb bandwidth, h = sd(x)*N^(-r)}

\item{cutoff.level}{a numeric number to generate an interior-point indicator}

\item{num.threads}{an integer number indicating the number of threads to be used. The default is
parallel::detectCores()-1L}

\item{...}{further arguments in maxLik::maxLik}
}
\value{
object of class 'maxLik'
}
\description{
This implements quasi maximum likelihood estimation using "maxLik::maxLik" to find identifiable parameters.
This runs the estimation twice, where the first estimation is the "pilot" version and the second
is the main one with interior points searched based on the "index" obtained from the pilot version.
}
