% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiBRM.R
\name{summary.semiBRM}
\alias{summary.semiBRM}
\title{Summary of quasi maximum likelihood estimation}
\usage{
\method{summary}{semiBRM}(object, ...)
}
\arguments{
\item{object}{a fitted 'semiBRM' object.}

\item{...}{further arguments (currently ignored).}
}
\description{
This shows summary of quasi maximum likelihood estimation of the parameters in the fitted
semiparametric binary response model.
}
