% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-func.R
\name{TrimmingIndicator}
\alias{TrimmingIndicator}
\title{Trimming indicator}
\usage{
TrimmingIndicator(x, f)
}
\arguments{
\item{x}{a numeric vector or matrix.}

\item{f}{a numeric number that indicates the cut-off percentile level.}
}
\value{
an integer vector consisting only of ones and zeros.

If \code{x} is a matrix, interior points are searched on each column, and
the element-wise product of the trimming indicators of the columns are returned.
}
\description{
This returns an integer vector whose i-th element takes 1L if the i-th element
of \code{x} lies between f*100 and (1-f)*100 -th percentiles, i.e.,
an interior point, and 0L otherwise.
}
\examples{
n <- 500L
x <- rnorm(n)
f <- 0.025

trimming <- TrimmingIndicator(x, f)
}
